fix: (#451) 自己发布的帖子评论置顶文案替换

+++ src/components/FeedCard/GroupFeedCard.vue
@@ -7,6 +7,7 @@ import {
   applyTopPostComment,
   deletePostComment
 } from "@/api/group.js";
+
 export default {
   name: "GroupFeedCard",
   extends: FeedCard,
@@ -120,24 +121,28 @@ export default {
       bus.$emit("actionSheet", [...base], "取消");
     },
     commentAction({ isMine = false, placeholder, reply_user, comment }) {
-      isMine
-        ? bus.$emit("actionSheet", [
-            {
-              text: "申请评论置顶",
-              method: () => {
-                bus.$emit("applyTop", {
-                  type: "postComment",
-                  api: applyTopPostComment,
-                  payload: { postId: this.feedID, commentId: comment.id }
-                });
-              }
-            },
-            { text: "删除评论", method: () => this.deleteComment(comment.id) }
-          ])
-        : this.handleComment({
-            placeholder,
-            reply_user
-          });
+      if (isMine) {
+        const isOwner = this.feed.user.id === this.CURRENTUSER.id;
+        bus.$emit("actionSheet", [
+          {
+            text: isOwner ? "评论置顶" : "申请评论置顶",
+            method: () => {
+              bus.$emit("applyTop", {
+                isOwner,
+                type: "postComment",
+                api: applyTopPostComment,
+                payload: { postId: this.feedID, commentId: comment.id }
+              });
+            }
+          },
+          { text: "删除评论", method: () => this.deleteComment(comment.id) }
+        ]);
+      } else {
+        this.handleComment({
+          placeholder,
+          reply_user
+        });
+      }
     },
     deleteComment(commentId) {
       deletePostComment(this.feedID, commentId).then(() => {
